requirements.txt
# Updated requirements with modern, compatible versions
# Core ML/DL frameworks
numpy>=1.21.0,<2.0.0
tensorflow>=2.13.0,<2.16.0

# Visualization
matplotlib>=3.5.0,<4.0.0

# Configuration file support
pyyaml>=6.0
jsonschema>=4.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# HTTP and utilities
requests>=2.28.0
certifi>=2023.5.0
idna>=3.4
urllib3>=1.26.0,<3.0.0

# Google Cloud / TensorBoard
tensorboard>=2.13.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
grpcio>=1.54.0
protobuf>=4.21.0,<5.0.0

# Date and time utilities  
python-dateutil>=2.8.2

# Async and crypto
pyasn1>=0.5.0
pyasn1-modules>=0.3.0
rsa>=4.9

# Other utilities
six>=1.16.0
Werkzeug>=2.3.0
wrapt>=1.15.0
oauthlib>=3.2.0
requests-oauthlib>=1.3.1
cachetools>=5.3.0
Markdown>=3.4.0

# DP-specific libraries
tensorflow-privacy>=0.9.0

# Note: Updated to TensorFlow 2.13+ for better security and performance
# Note: Added configuration file support (YAML/JSON)
# Note: Added testing frameworks
# Note: Added tensorflow-privacy for DP mechanisms